+++ src/main/java/location/domain/Triangle.java
@@ -26,7 +26,7 @@ public class Triangle extends Shape {
     }
 
     private double calculateTilt(final Point point, final Point point1) {
-        return (point.getY()-point1.getY())/(point.getX()-point1.getX());
+        return (double)(point.getY()-point1.getY())/(double)(point.getX()-point1.getX());
     }
